Add some colors to the default highlight groups
It adds/rewrites colors to some default highlights groups like ErrorMsg
or WarningMsg or Title. The intention is to cover all the default groups
before, maybe, trying to cover 'filetype' related groups.

It must be noted that the color named error and warning are used in the
interface, but, only to put the emphasize on warning or error messages.