Add almost all the default highlighting groups
Almost all the default highligting groups have been added. Those which
remain commented are related to the GUI, and they will be added later.
The color of some of the already existing groups has changed, but it has
not been logged because nothing is fixed right now.

The function s:HL has also been change a little bit in order to support
an optional third argument to provide the value of 'gui'.