Modify the colors of the PMenu
The colors of the scrollbar and its thumb which are both associated to
the Pmenu have been modified after using it on an example.

The scrollbar has now the same color as the main part of the PMenu, and
the color of its thumb is the one of the 'linenumber'.