Add the initialisation step
The initialisation step consists in cleaning the highlights and the
syntax from previous schemes, and setting the background accordingly.

Plus, the variables used within the script are also initialized.

It must be noted that the colorscheme is only defined for the GUI and
for versions of Vim that has the 'termguicolors' option set. In addition
the version must be greater than '580'.