Remove the .gitignore file
The .gitignore file was only containing the line '*.swp', which are
files related to Vim. Given that it focuses to much on a given editor
and not on files or repertories of the project, the line has been
removed.

In the same time, the entire file is also removed given that it is
empty.