Add the support for 8-bit colors
The true color and its 8-bit association are managed in the same tuples.
In other words, s:colors is now an array of tuples.

Rigth now, I think the mapping between the true and the 8-bit colors can
be improved. I have just used an approximation script I found on the
internet (URL too long).