Allow the highlighting of the Javadoc
It adds a mechanism that allows to trigger the highlight of the Javadoc.
In order to do so, the user must add the following to its .vimrc file
before declaring the colorscheme:

    let g:odyssey_java_javadoc = 1
    colorscheme odyssey

It must be noted that if g:odyssey_java_javadoc equals 0, the
highlighting of the Javadoc is not taken into account.